+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Sun Nov 27 14:49:22 BRT 2022
+gradle.version=7.4
