+++ src/main/java/project/kazumy/realhosting/discord/commands/base/BaseSlashCommand.java
@@ -22,4 +22,8 @@ public abstract class BaseSlashCommand {
     public void register() {
         InitBot.commandManager.addSlashCommand(this);
     }
+
+    public InitBot initBotInstance() {
+        return InitBot.getInstance();
+    }
 }
