+++ src/main/java/project/kazumy/realhosting/discord/commands/prefix/ForceApprovalCommand.java
@@ -1,4 +1,4 @@
-package project.kazumy.realhosting.discord.commands;
+package project.kazumy.realhosting.discord.commands.prefix;
 
 import lombok.val;
 import net.dv8tion.jda.api.EmbedBuilder;
