+++ src/main/java/project/kazumy/realhosting/discord/listener/InteractionService.java
@@ -2,6 +2,7 @@ package project.kazumy.realhosting.discord.listener;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import project.kazumy.realhosting.discord.InitBot;
 
 @Data
 @AllArgsConstructor
@@ -10,4 +11,8 @@ public abstract class InteractionService<T> {
     private final String id;
 
     public abstract void execute(T event);
+
+    public InitBot initBotInstance() {
+        return InitBot.getInstance();
+    }
 }
