+++ src/main/java/project/kazumy/realhosting/discord/services/panel/ServerType.java
@@ -25,6 +25,24 @@ public enum ServerType {
                             .retrieveNestById(1).execute(), 3).execute(),
             bungeecordEnviroment()),
 
+    POCKET_MINE("ghcr.io/parkervcp/yolks:debian",
+            InitBot.panelManager.getApplication()
+                    .retrieveEggById(InitBot.panelManager.getApplication()
+                            .retrieveNestById(1).execute(), 23).execute(),
+            pocketmineEnviroment()),
+
+    NUKKIT("ghcr.io/pterodactyl/yolks:java_17",
+            InitBot.panelManager.getApplication()
+                    .retrieveEggById(InitBot.panelManager.getApplication()
+                            .retrieveNestById(1).execute(), 22).execute(),
+            nukkitEnviroment()),
+
+    BEDROCK("ghcr.io/parkervcp/yolks:debian",
+            InitBot.panelManager.getApplication()
+                    .retrieveEggById(InitBot.panelManager.getApplication()
+                            .retrieveNestById(1).execute(), 21).execute(),
+            bedrockEnviroment()),
+
     JAVA_BOT("ghcr.io/parkervcp/yolks:java_17",
             InitBot.panelManager.getApplication()
                     .retrieveEggById(InitBot.panelManager.getApplication()
@@ -48,6 +66,31 @@ public enum ServerType {
         Map<String, EnvironmentValue<?>> map = new HashMap<>();
         map.put("SERVER_JARFILE", EnvironmentValue.of("bungeecord.jar"));
         map.put("BUNGEE_VERSION", EnvironmentValue.of("latest"));
+        return map;
+    }
+
+    public static Map<String, EnvironmentValue<?>> pocketmineEnviroment() {
+        Map<String, EnvironmentValue<?>> map = new HashMap<>();
+        map.put("VERSION", EnvironmentValue.of("pm4"));
+        return map;
+    }
+
+    public static Map<String, EnvironmentValue<?>> nukkitEnviroment() {
+        Map<String, EnvironmentValue<?>> map = new HashMap<>();
+        map.put("SERVER_JARFILE", EnvironmentValue.of("server.jar"));
+        map.put("NUKKIT_VERSION", EnvironmentValue.of("latest"));
+        map.put("DL_PATH", EnvironmentValue.of(""));
+        return map;
+    }
+
+    public static Map<String, EnvironmentValue<?>> bedrockEnviroment() {
+        Map<String, EnvironmentValue<?>> map = new HashMap<>();
+        map.put("BEDROCK_VERSION", EnvironmentValue.of("latest"));
+        map.put("LD_LIBRARY_PATH", EnvironmentValue.of("."));
+        map.put("SERVERNAME", EnvironmentValue.of("Bedrock Dedicated Server"));
+        map.put("GAMEMODE", EnvironmentValue.of("survival"));
+        map.put("DIFFICULTY", EnvironmentValue.of("easy"));
+        map.put("CHEATS", EnvironmentValue.of("false"));
         return map;
     }
