+++ src/test/java/MainRealTest.java
@@ -1,7 +1,14 @@
+import com.mercadopago.MercadoPagoConfig;
+import com.mercadopago.client.payment.PaymentClient;
+import com.mercadopago.net.MPSearchRequest;
+import lombok.SneakyThrows;
+import lombok.val;
 
+import java.math.BigDecimal;
 
 public class MainRealTest {
 
+    @SneakyThrows
     public static void main(String[] args) {
 
     }
